refactor shaps; move TShapPreparedTrees to new file
