+++ catboost/libs/fstr/util.h
@@ -13,6 +13,9 @@ TVector<double> CollectLeavesStatistics(
     const TFullModel& model,
     NPar::TLocalExecutor* localExecutor);
 
-bool TryGetLossDescription(const TFullModel& model, NCatboostOptions::TLossDescription& lossDescription);
+bool TryGetLossDescription(const TFullModel& model, NCatboostOptions::TLossDescription* lossDescription);
+
+bool TryGetObjectiveMetric(const TFullModel& model, NCatboostOptions::TLossDescription* lossDescription);
 
 void CheckNonZeroApproxForZeroWeightLeaf(const TFullModel& model);
+
